requirements.txt
langchain
openai
pymysql
azure-functions
